+++ src/ECS/Types.h
@@ -1,19 +1,18 @@
 #pragma once
 
 #include "Draug.h"
-#include "Core/Types.h"
 
 namespace Draug {
 namespace ECS {
 namespace details {
-const uint32 ENTITY_INDEX_BITS = 32;
-const uint64 ENTITY_INDEX_MASK = (uint64(1) << ENTITY_INDEX_BITS) - 1;
-const uint64 ENTITY_GENERATION_MASK = (uint64(1) << ENTITY_INDEX_BITS) - 1;
+const unsigned int ENTITY_INDEX_BITS = 32;
+const size_t ENTITY_INDEX_MASK = (size_t(1) << ENTITY_INDEX_BITS) - 1;
+const size_t ENTITY_GENERATION_MASK = (size_t(1) << ENTITY_INDEX_BITS) - 1;
 
-const uint32 MAX_COMPONENTS = 64;
+const uint16_t MAX_COMPONENTS = 64;
 }
-typedef uint64 EntityId;
-typedef uint32 ComponentId;
+typedef size_t EntityId;
+typedef size_t ComponentId;
 
 typedef std::bitset<details::MAX_COMPONENTS> ComponentMask;
 }
